requirements.txt
python_binance
TA-Lib
numpy
pandas
websocket_client==0.40.0
matplotlib.pyplot
